Rename profilePass to accProfPass to differentiate from future pass commits. Move dummy apps from profilePass to new dummyApps directory. Rename some scripts and add comments. Rename attester_sw library to accproff_soft. Update README files to reflect the latest changes

+++ scripts/sel4-compile.sh
@@ -1,7 +1,20 @@
 #!/bin/bash
 
+# $1: The relative path to the LLVM pass used. 
+
+# $2: The relative path to the text file containting the list of functions to be
+# profiled.
+
+if [[ $# -lt 2 ]] ; then
+	echo 'No arguments provided!'
+	echo -e '\t$1: The relative path to the LLVM pass used.'
+	echo -e '\t$2: The relative path to the text file containting the list of'
+	echo -e '\t    functions to be profiled.'
+	exit 0
+fi
+
 export LLVM_DIR=/usr/lib/llvm-15
 export LLVM_SYMBOLIZER_PATH=$LLVM_DIR/bin/llvm-symbolizer
 rm -rf ./*
-../init-build.sh -DPLATFORM=zcu102 -DTRIPLE=aarch64-linux-gnu
+../init-build.sh -DPLATFORM=zcu102 -DTRIPLE=aarch64-linux-gnu -DLLVMPass=$1 -DFunctionsFile=$2
 ninja
